feat: :seedling: Correction User

+++ app/Http/Controllers/UsuarioController.php
@@ -2,31 +2,59 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\UsuarioRequest; // Presumindo que há uma classe de requisição para validação
+use App\Http\Resources\UsuarioResource;
+use App\Models\Usuario;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
-class UsuarioController extends Controller {
-    public function index() {
-        $usuarios = Usuario::all();
-        return response()->json($usuarios);
+class UsuarioController extends Controller
+{
+    public function index()
+    {
+        return UsuarioResource::collection(Usuario::all());
     }
 
-    public function store(Request $request) {
-        $usuario = new Usuario($request->nome, $request->email, $request->senha, $request->perfil_id);
-        $usuario->save();
-        return response()->json($usuario);
+    
+    public function store(UsuarioRequest $request)
+    {
+        $usuario = Usuario::create([
+            'nome' => $request->nome,
+            'email' => $request->email,
+            'senha' => Hash::make($request->senha),
+            'perfil_id' => $request->perfil_id,
+        ]);
+
+        $resource = new UsuarioResource($usuario);
+        return $resource->response()->setStatusCode(201); 
     }
 
-    public function update(Request $request, $id) {
+    public function update(UsuarioRequest $request, string $id)
+    {
         $usuario = Usuario::find($id);
-        $usuario->nome = $request->nome;
-        $usuario->email = $request->email;
-        $usuario->save();
-        return response()->json($usuario);
+
+        if (!$usuario) {
+            return response(['error' => 'Usuário não encontrado.'], 404);
+        }
+
+        $usuario->update([
+            'nome' => $request->nome,
+            'email' => $request->email,
+        ]);
+
+        $resource = new UsuarioResource($usuario);
+        return $resource->response()->setStatusCode(200); 
     }
 
-    public function destroy($id) {
+    public function destroy(string $id)
+    {
         $usuario = Usuario::find($id);
+
+        if (!$usuario) {
+            return response(['error' => 'Usuário não encontrado.'], 404);
+        }
+
         $usuario->delete();
-        return response()->json('Usuario deletado');
+        return response(['message' => 'Usuário deletado com sucesso.'], 200);
     }
 }
